backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for Todo Platform — Phase II"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.0",
    "pydantic[email]>=2.9.0",
    "pyjwt[cryptography]>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "--cov=. --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*", ".venv/*"]
